+++ src/check_datapackage/check.py
@@ -51,17 +51,56 @@ class DataPackageError(Exception):
         self,
         issues: list[Issue],
     ) -> None:
-        """Create the DataPackageError attributes from issues."""
-        # TODO: Switch to using `explain()` once implemented
-        errors: list[str] = _map(
-            issues,
-            lambda issue: f"- Property `{issue.jsonpath}`: {issue.message}\n",
+        """Create the DataPackageError from issues."""
+        super().__init__(explain(issues))
+
+
+def explain(issues: list[Issue]) -> str:
+    """Explain the issues in a human-readable format.
+
+    Args:
+        issues: A list of `Issue` objects to explain.
+
+    Returns:
+        A human-readable explanation of the issues.
+
+    Examples:
+        ```{python}
+        import check_datapackage as cdp
+
+        issue = cdp.Issue(
+            jsonpath="$.resources[2].title",
+            type="required",
+            message="The `title` field is required but missing at the given JSON path.",
         )
-        message: str = (
-            "There were some issues found in your `datapackage.json`:\n\n"
-            + "\n".join(errors)
-        )
-        super().__init__(message)
+
+        cdp.explain([issue])
+        ```
+    """
+    issue_explanations: list[str] = _map(
+        issues,
+        _create_explanation,
+    )
+    num_issues = len(issue_explanations)
+    singular_or_plural = " was" if num_issues == 1 else "s were"
+    return (
+        f"{num_issues} issue{singular_or_plural} found in your `datapackage.json`:\n\n"
+        + "\n".join(issue_explanations)
+    )
+
+
+def _create_explanation(issue: Issue) -> str:
+    """Create an informative explanation of what went wrong in each issue."""
+    # Remove suffix '$' to account for root path when `[]` is passed to `check()`
+    property_name = issue.jsonpath.removesuffix("$").split(".")[-1]
+    number_of_carets = len(str(issue.instance))
+    return (  # noqa: F401
+        f"At package{issue.jsonpath.removeprefix('$')}:\n"
+        "|\n"
+        f"| {property_name}{': ' if property_name else '  '}{issue.instance}\n"
+        f"| {' ' * len(property_name)}  {'^' * number_of_carets}\n"
+        f"{issue.message}\n"
+    )
 
 
 def check(
@@ -350,11 +389,14 @@ def _handle_S_resources_x_schema_fields_x(
 
     E.g., if the field type is `string`, flag errors for the string-based schema only.
     """
-    edits = SchemaErrorEdits()
+    edits = SchemaErrorEdits(remove=[parent_error])
     errors_in_group = _get_errors_in_group(schema_errors, parent_error)
-    edits.remove.append(parent_error)
 
-    field_type: str = parent_error.instance.get("type", "string")
+    parent_instance = parent_error.instance
+    if not isinstance(parent_instance, dict):
+        return edits
+
+    field_type: str = parent_instance.get("type", "string")
 
     # The field's type is unknown
     if field_type not in FIELD_TYPES:
@@ -366,7 +408,7 @@ def _handle_S_resources_x_schema_fields_x(
             type="enum",
             jsonpath=f"{parent_error.jsonpath}.type",
             schema_path=parent_error.schema_path,
-            instance=parent_error.instance,
+            instance=parent_instance,
         )
         # Replace all errors with an unknown field error
         edits.add.append(unknown_field_error)
@@ -489,11 +531,14 @@ def _handle_S_resources_x_schema_foreign_keys(
     edits = SchemaErrorEdits(remove=[parent_error])
     errors_in_group = _get_errors_in_group(schema_errors, parent_error)
 
-    key = parent_error.instance.get("fields", None)
-    key_type = type(key)
+    parent_instance = parent_error.instance
+    key_exists = isinstance(parent_instance, dict) and "fields" in parent_instance
 
     # If the key type is correct, use that schema
-    if key_type in FOREIGN_KEY_TYPES:
+    if (
+        key_exists
+        and (key_type := type(parent_instance["fields"])) in FOREIGN_KEY_TYPES
+    ):
         schema_part = f"foreignKeys/items/oneOf/{FOREIGN_KEY_TYPES.index(key_type)}/"
         edits.remove.extend(
             _filter(
@@ -512,7 +557,7 @@ def _handle_S_resources_x_schema_foreign_keys(
     edits.remove.extend(key_type_errors)
 
     # If the key exists, flag incorrect type
-    if key is not None:
+    if key_exists:
         edits.add.append(
             SchemaError(
                 message=(
@@ -655,6 +700,7 @@ def _create_issue(error: SchemaError) -> Issue:
         message=error.message,
         jsonpath=error.jsonpath,
         type=error.type,
+        instance=error.instance,
     )