+++ CHANGELOG.md
@@ -9,6 +9,18 @@ often, sometimes several in a day. It also means any individual release
 will not have many changes within it. Below is a list of releases along
 with what was changed within it.
 
+## 0.23.1 (2025-11-27)
+
+### Fix
+
+- :bug: add dict check when getting parent instance (#214)
+
+## 0.23.0 (2025-11-27)
+
+### Feat
+
+- :sparkles: `explain()` errors (#208)
+
 ## 0.22.2 (2025-11-21)
 
 ### Fix